Use an explicit authenticating realm order instead of the dynamic one
Also disable binding of the legacy LDAP and MongoDB realm.s
